fix: fluid amount can't be showing correctly in TankBackpack

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/entity/backpack/data/TankBackpackData.java
@@ -66,7 +66,7 @@ public class TankBackpackData extends SimpleContainer implements IBackpackData {
             this.tankFluidCount = tank.amount;
             // amount改变时发包同步客户端流体amount
             if (TankBackpackData.this.maid.getOwner() instanceof ServerPlayer serverPlayer) {
-                ServerPlayNetworking.send(serverPlayer, new SyncFluidAmountPackage((int) this.tankFluidCount));
+                ServerPlayNetworking.send(serverPlayer, new SyncFluidAmountPackage(this.tankFluidCount));
             }
             ResourceLocation key = BuiltInRegistries.FLUID.getKey(tank.getResource().getFluid());
             maid.setBackpackFluid(key.toString());
